Compare dates in modify_data method
In the `modify_data` method of the import script, added a date comparison check to determine if a file has been modified or added since the last import to OpenSearch. The file's timestamp is compared with the provided last import timestamp to make the decision.Committing this change to ensure accurate handling of file modifications during the import process.

Co-authored-by: leondaniel22 <leondaniel22@web.de>
Signed-off-by:  Amir-Hussein-OTH <miro.apple94@gmail.com>