modified the code
I have added an import_all parameter to the OpenSearchManager class constructor.
This parameter is used to determine whether to retry importing failed entries or not.
If import_all is set to True, the failed entries will be retried individually or in smaller batches using the retry_import() method.
If import_all is set to False, the failed entries will be logged or handled separately for further analysis.

Co-authored-by: leondaniel22 <leondaniel22@web.de>
Signed-off-by: Amir-Hussein-OTH <miro.apple94@gmail.com>